:sparkles: feat: 触摸设置迁移到角色编辑
