+++ src/components/agent/AgentMeta/index.tsx
@@ -1,21 +1,24 @@
 import { Avatar } from '@lobehub/ui';
 import { Typography } from 'antd';
+import React from 'react';
 
 import { AgentMeta } from '@/types/agent';
 
 import { useStyles } from './style';
 
 interface AgentMetaProps {
+  className?: string;
   meta?: AgentMeta;
+  style?: React.CSSProperties;
 }
 
 export default (props: AgentMetaProps) => {
-  const { styles } = useStyles();
-  const { meta } = props;
+  const { styles, cx } = useStyles();
+  const { meta, style, className } = props;
   const { avatar, name, description } = meta || {};
 
   return (
-    <div className={styles.container}>
+    <div className={cx(styles.container, className)} style={style}>
       <Avatar avatar={avatar} size={36} />
       <div className={styles.content}>
         <div className={styles.title}>{name}</div>
