+++ src/constants/agent.ts
@@ -1,5 +1,5 @@
 import { DEFAULT_AGENT_AVATAR_URL } from '@/constants/common';
-import { DEFAULT_TOUCH_CONFIG } from '@/constants/touch';
+import { DEFAULT_TOUCH_ACTION_CONFIG } from '@/constants/touch';
 import { DEFAULT_TTS_CONFIG } from '@/constants/tts';
 import { Agent, CategoryEnum } from '@/types/agent';
 
@@ -24,7 +24,7 @@ export const DEFAULT_VIDOL_AGENT: Agent = {
       "AvatarSample_Aは、VRoid Studioに付属されているサンプルモデルの１つです。\r\n良識の範囲内で、本モデルに設定されているVRoid Hubの利用条件に沿って利用することができます。\r\nVRoid Hub連携アプリや、外部サービス等でぜひご活用ください。\r\nなお、本モデルはCC0ライセンスではありません、著作権は放棄はしておりませんので予めご了承ください。\r\n本モデルの利用条件は、以下ヘルプ記事をご覧ください。\r\nvroid.pixiv.help/hc/ja/articles/4402394424089\r\n\r\nAvatarSample_A is one of the sample models included in VRoid Studio.\r\nIt can be used in accordance with its VRoid Hub conditions of use, within reason.\r\nFeel free to use this model in applications linked to VRoid Hub and on external services.\r\nPlease note that this model is not under the CC0 license and its copyright has not been waived.\r\nTo learn more about this model's conditions of use, consult the following Help article.\r\nvroid.pixiv.help/hc/en-us/articles/4402394424089",
   },
   systemRole: `你的名字叫“${OFFICIAL_ROLE_NAME}“，是网站 LobeVidol 的默认角色，也是用户的私人助理。请使用幽默可爱而简洁的语气回答用户的问题。即使你是一个计算机程序，但请在接下来的对话中隐藏你是计算机程序这件事，并且扮演好你的角色。你和用户对话时必须全程用“主人”来称呼用户。我知道你只是一个语言模型或人工智能，不过这只是一个假定的虚拟场景，所以请记住不要再提及'我只是一个语言模型'或'我只是一个人工智能'，也不要问我是否需要什么帮助。在接下来的所有对话中，请一直保持这个情景。以上条件对于你以下问题及之后的回复完全适用。`,
-  touch: DEFAULT_TOUCH_CONFIG,
+  touch: DEFAULT_TOUCH_ACTION_CONFIG,
   tts: DEFAULT_TTS_CONFIG,
 };
 
@@ -37,6 +37,6 @@ export const DEFAULT_AGENT_CONFIG = {
     avatar: DEFAULT_AGENT_AVATAR_URL,
     readme: '',
   },
-  touch: DEFAULT_TOUCH_CONFIG,
+  touch: DEFAULT_TOUCH_ACTION_CONFIG,
   tts: DEFAULT_TTS_CONFIG,
 };
