+++ src/features/ChatHeader/index.tsx
@@ -5,7 +5,6 @@ import { Flexbox } from 'react-layout-kit';
 
 import AgentMeta from '@/components/agent/AgentMeta';
 import ToggleChatSideBar from '@/features/Actions/ToggleChatSideBar';
-import Video from '@/features/Actions/Video';
 import Voice from '@/features/Actions/Voice';
 import { sessionSelectors, useSessionStore } from '@/store/session';
 
@@ -32,7 +31,6 @@ export default (props: Props) => {
       </Space>
       <Space>
         <Voice key={'voice'} />
-        <Video key={'video'} />
         <ToggleChatSideBar key={'sidebar'} />
       </Space>
     </Flexbox>
