+++ src/panels/RolePanel/RoleEdit/Model/Touch/index.tsx
@@ -1,6 +1,8 @@
 import { createStyles } from 'antd-style';
 import classNames from 'classnames';
-import React, { memo } from 'react';
+import React, { memo, useState } from 'react';
+
+import { TouchAreaEnum } from '@/types/touch';
 
 import ActionList from './ActionList';
 import SideBar from './SideBar';
@@ -27,11 +29,12 @@ interface TouchProps {
 const Touch = (props: TouchProps) => {
   const { style, className } = props;
   const { styles } = useStyles();
+  const [currentTouchArea, setCurrentTouchArea] = useState(TouchAreaEnum.Head);
 
   return (
     <div className={classNames(className, styles.container)} style={style}>
-      <SideBar />
-      <ActionList />
+      <SideBar currentTouchArea={currentTouchArea} setCurrentTouchArea={setCurrentTouchArea} />
+      <ActionList currentTouchArea={currentTouchArea} />
     </div>
   );
 };
