+++ src/store/agent/index.ts
@@ -8,6 +8,7 @@ import { StateCreator } from 'zustand/vanilla';
 
 import { DEFAULT_AGENT_CONFIG, LOBE_VIDOL_DEFAULT_AGENT_ID } from '@/constants/agent';
 import { Agent, AgentMeta } from '@/types/agent';
+import { TouchActionConfig } from '@/types/touch';
 import { TTS } from '@/types/tts';
 import { mergeWithUndefined } from '@/utils/common';
 
@@ -71,6 +72,10 @@ export interface AgentStore {
    * 更新角色 TTS
    */
   updateAgentTTS: (tts: DeepPartial<TTS>) => void;
+  /**
+   * 更新角色触摸配置
+   */
+  updateAgentTouch: (touch: DeepPartial<TouchActionConfig>) => void;
 }
 
 const createAgentStore: StateCreator<AgentStore, [['zustand/devtools', never]]> = (set, get) => ({
@@ -130,6 +135,11 @@ const createAgentStore: StateCreator<AgentStore, [['zustand/devtools', never]]> 
   updateAgentMeta: (meta) => {
     const { updateAgentConfig } = get();
     updateAgentConfig({ meta });
+  },
+
+  updateAgentTouch: (touch) => {
+    const { updateAgentConfig } = get();
+    updateAgentConfig({ touch });
   },
   updateAgentTTS: (tts) => {
     const { updateAgentConfig } = get();
